prepped for frontent adaptions for Project & Tags

+++ PersonalSite.Api/Application/Tags/DeleteTag/DeleteTagCommandHandler.cs
@@ -1,4 +1,5 @@
 using PersonalSite.Api.Domain.Actors;
+using PersonalSite.Api.Domain.Exceptions;
 using PersonalSite.Api.Domain.Tags;
 using PersonalSite.Api.Storage.Tags;
 
@@ -24,6 +25,17 @@ public sealed class DeleteTagCommandHandler(
         TagPermissions.EnsureCanManage(
             actor,
             tag);
+
+        var isInUse =
+            await tagRepository.IsInUseAsync(
+                id,
+                cancellationToken);
+
+        if (isInUse)
+        {
+            throw new TagInUseException(
+                "This tag cannot be deleted because it is used by a project.");
+        }
 
         await tagRepository.DeleteAsync(
             tag,
